requirements.txt
asn1crypto==0.24.0
certifi==2019.6.16
chardet==3.0.4
cryptography==2.1.4
enum34==1.1.6
idna==2.8
ipaddress==1.0.17
keyring==10.6.0
keyrings.alt==3.0
pycrypto==2.6.1
pygobject==3.26.1
python-dotenv==0.10.3
pyxdg==0.25
requests==2.22.0
SecretStorage==2.3.1
six==1.11.0
spotipy==2.4.4
tinytag==1.2.2
typing==3.7.4
urllib3==1.25.3
